+++ src/command/dev/mod.rs
@@ -20,9 +20,6 @@ mod do_dev;
 mod introspect;
 
 #[cfg(feature = "composition-js")]
-mod protocol;
-
-#[cfg(feature = "composition-js")]
 mod router;
 
 #[cfg(feature = "composition-js")]
@@ -33,6 +30,9 @@ mod netstat;
 
 #[cfg(not(feature = "composition-js"))]
 mod no_dev;
+
+#[cfg(feature = "composition-js")]
+mod types;
 
 #[cfg(feature = "composition-js")]
 mod watcher;
