+++ src/command/dev/schema.rs
@@ -9,8 +9,7 @@ use rover_client::blocking::StudioClient;
 use crate::options::ProfileOpt;
 use crate::{
     command::dev::{
-        netstat::normalize_loopback_urls, protocol::FollowerMessenger,
-        watcher::SubgraphSchemaWatcher, SupergraphOpts,
+        netstat::normalize_loopback_urls, watcher::SubgraphSchemaWatcher, SupergraphOpts,
     },
     options::OptionalSubgraphOpts,
     utils::client::StudioClientConfig,
